Moves tests & methods around to get shift class passing tests and finished.
